refactor, add more doc, update to larger ODE problem
