+++ mEncoder/encoder.py
@@ -2,18 +2,32 @@
 Materials for a simple linear encoder, and its analytical reverse.
 """
 
-import theano.tensor as T
+import theano.tensor as tt
 import theano
 import numpy as np
 from typing import List
 
 TheanoFunction = theano.compile.function_module.Function
 
-class dA:
-    """A simple linear autoencoder. """
 
-    def __init__(self, input_data, n_hidden=1, n_params=12,
-                 par_modulation_scale=4):
+class AutoEncoder:
+    """
+    A simple linear autoencoder.
+
+    :param input_data:
+        input data for the encoder
+
+    :param n_hidden:
+        number of latent variables
+
+    :param n_params:
+        number of parameters that to which the embedding will be inflated to
+    """
+
+    def __init__(self,
+                 input_data: np.ndarray,
+                 n_hidden: int = 1,
+                 n_params: int = 12):
         self.n_visible = input_data.shape[1]
         assert n_hidden < self.n_visible
         assert n_hidden <= n_params
@@ -27,10 +41,10 @@ class dA:
         self.n_encoder_pars = self.n_encode_weights + \
             self.n_inflate_weights + self.n_inflate_bias
 
-        self.encoder_pars = T.specify_shape(T.vector('encoder_pars'),
-                                            (self.n_encoder_pars,))
+        self.encoder_pars = tt.specify_shape(tt.vector('encoder_pars'),
+                                             (self.n_encoder_pars,))
 
-        self.par_modulation_scale = par_modulation_scale
+        # self.par_modulation_scale = par_modulation_scale
 
         self.x_names = [
             f'ecoder_{iw}_weight' for iw in range(self.n_encode_weights)
@@ -40,31 +54,53 @@ class dA:
             f'inflate_{iw}_bias' for iw in range(self.n_inflate_bias)
         ]
 
-    def encode(self, pIn):
-        """ Run the input through the encoder. """
-        W = T.reshape(pIn[0:self.n_encode_weights],
-                      (self.n_visible, self.n_hidden))
-        return T.dot(self.data, W)
+    def encode(self, parameters: tt.vector):
+        """
+        Run the input through the encoder.
 
-    def inflate_params(self, embedded_data, pIn):
-        """ Inflate the input to parameters. """
-        W_p = T.reshape(pIn[self.n_encode_weights:
-                            self.n_encode_weights+self.n_inflate_weights],
-                        (self.n_hidden, self.n_params))
+        :param parameters:
+            parametrization of full autoencoder
+        """
+        W = tt.reshape(parameters[0:self.n_encode_weights],
+                       (self.n_visible, self.n_hidden))
+        return tt.dot(self.data, W)
+
+    def inflate_params(self, embedded_data, parameters):
+        """
+        Inflate the input to parameters.
+
+        :param parameters:
+            parametrization of full autoencoder
+        """
+        W_p = tt.reshape(
+            parameters[self.n_encode_weights:
+                       self.n_encode_weights+self.n_inflate_weights],
+            (self.n_hidden, self.n_params)
+        )
         # bias = pIn[-self.n_inflate_bias:]
-        return T.dot(embedded_data, W_p)
+        return tt.dot(embedded_data, W_p)
         #return T.nnet.sigmoid(T.dot(embedded_data, W_p) + bias) \
         #    * self.par_modulation_scale*2 - self.par_modulation_scale
 
-    def encode_params(self, pIn):
-        """ Run the encoder and then inflate to parameters. """
-        return self.inflate_params(self.encode(pIn), pIn)
+    def encode_params(self, parameters: tt.vector):
+        """
+        Run the encoder and then inflate to parameters.
+
+        :param parameters:
+            parametrization of full autoencoder
+        """
+        return self.inflate_params(self.encode(parameters), parameters)
     
-    def decode(self, embedded_data, pIn):
-        """ Run the input through the analytical decoder. """
-        W = T.reshape(pIn[0:self.n_encode_weights],
-                      (self.n_visible, self.n_hidden))
-        return T.dot(embedded_data, T.nlinalg.pinv(W))
+    def decode(self, embedded_data, parameters: tt.vector):
+        """
+        Run the input through the analytical decoder.
+
+        :param parameters:
+            parametrization of full autoencoder
+        """
+        W = tt.reshape(parameters[0:self.n_encode_weights],
+                       (self.n_visible, self.n_hidden))
+        return tt.dot(embedded_data, tt.nlinalg.pinv(W))
 
     def compile_inflate_pars(self) -> TheanoFunction:
         """
